Return error if already has a product with that id and delete image

+++ src/controllers/productController.js
@@ -1,3 +1,4 @@
+import ErrorHandler from "../middlewares/error/errorHandler.js";
 import Product from "../models/productModel.js";
 import catchAsyncError from "../utils/catchAsyncError.js";
 import deleteFile from "../utils/files/deleteFile.js";
@@ -8,10 +9,20 @@ export const createProduct = catchAsyncError(async (req, res, next) => {
   // parsing the file url to save database with the subfoldername
   const fileUrl = fileUrlParser(req.file);
 
+  // finding the product with the product id if it exist or not
+  const existingProductId = await Product.findOne({
+    productId: req.body.productId,
+  });
+
+  // returning an error response if there is product with that id
+  if (existingProductId) {
+    deleteFile("products", req.file.filename);
+    return next(new ErrorHandler(409, "Product With That Id Already Exists"));
+  }
+
   // storing the product on database
   const product = await Product.create({ ...req.body, photo: fileUrl });
 
-  deleteFile("products", req.file.filename);
   res.status(200).json({
     success: true,
     product,
